requirements.txt
azure-cognitiveservices-speech==1.30.0
